refactor(download): move the acquisition cron bodies and their subscriber into plugins/download
`SearchMissing` and `RssSync` keep their registration, their cron expressions, their manual trigger
and their command audit trail in core; only their bodies move, along with the per-candidate indexer
fan-out and the RSS matchers. The scheduler delegates. Core keeps metadata refresh, the subtitle
jobs, rescan, markers and sprites.

The domain-event subscriber moves too. The bus itself stays in core with its event names unchanged —
what crossed the boundary is the subscribing side, not `emitDomain`/`onDomain`.

`AcquisitionCandidatesService` stays in core: four QueryBuilders over core tables, one interpolating
a raw SQL fragment naming core columns. The plugin calls it rather than owning it.

Verified by booting a real Nest application context from a fresh compile, not only by typechecking:
the previous PR in this series compiled clean, passed every unit test and then failed to start on an
import cycle, because nothing in the repo instantiates the module graph. This one resolves.

Refs: #894
